When board is viewed, nickname will be placed in writer

+++ src/main/java/com/example/prj1be/controller/MemberController.java
@@ -71,9 +71,17 @@ public class MemberController {
     }
 
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
+    public ResponseEntity<Member> view(String id, @SessionAttribute(value="login", required = false) Member login) {
         // TODO : loged in? -> if not, 401
         // TODO : authentication done? -> if not, 403
+
+        if(login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if(!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         Member member = service.getMember(id);
 
@@ -81,9 +89,17 @@ public class MemberController {
     }
 
     @DeleteMapping()
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id, @SessionAttribute(value = "login", required = false) Member login) {
         // TODO : loged in? -> if not, 401
         // TODO : authentication done? -> if not, 403
+
+        if(login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if(!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if(service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -92,8 +108,16 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
+    public ResponseEntity edit(@RequestBody Member member, @SessionAttribute(value = "login", required = false) Member login) {
         // TODO: check whether login / authentication done
+
+        if(login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if(!service.hasAccess(member.getId(), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if(service.update(member)) {
             return ResponseEntity.ok().build();
